requirements.txt
# Core orchestration
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0

# API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Data validation
pydantic>=2.8.0
pydantic-settings>=2.4.0

# LLM (Claude como alternativa justificada a Vertex AI Gemini)
anthropic>=0.34.0

# Observability
langfuse>=2.0.0

# RAG - Embeddings y Vector Store
sentence-transformers>=3.0.0
chromadb>=0.5.0

# GraphRAG
networkx>=3.3

# Hybrid search (BM25 sparse)
rank-bm25>=0.2.2

# PDF processing
pypdf>=4.3.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
httpx>=0.27.0

# Utils
python-dotenv>=1.0.0
